Suwat CR: remove BaseGitRepository abstract class and have a _legacyGitExeRepository in LibGit2Sharp for fallback
