+++ Kudu.Core/Kudu.Core.csproj
@@ -107,7 +107,6 @@
     <Compile Include="Settings\DiagnosticsSettingsManager.cs" />
     <Compile Include="SiteExtensions\SemanticVersion.cs" />
     <Compile Include="SiteExtensions\SiteExtensionManager.cs" />
-    <Compile Include="SourceControl\Git\BaseGitRepository.cs" />
     <Compile Include="SourceControl\Git\LibGit2SharpRepository.cs" />
     <Compile Include="Tracing\KuduEventSource.cs" />
     <Compile Include="Tracing\KuduSiteExtensionLogEvent.cs" />
